Split out receiveThemes
